Refactor diff logic into a Util class.
